requirements.txt
torch
transformers
peft
trl
bitsandbytes
datasets
faiss-gpu
sentence-transformers
pandas
wikipedia
tqdm